backend/manager_agent/requirements.txt
crewai
crewai-tools
python-dotenv
pandas
pyarrow
